--- DSharpPlus.Voice/Protocol/Dave/V1/Gateway/Payloads/IVoicePayload.cs
@@ -1,2 +1,0 @@
-namespace DSharpPlus.Voice.Protocol.Dave.V1.Gateway.Payloads;
-internal interface IVoicePayload;
