--- DSharpPlus.Voice/Protocol/Dave/V1/Gateway/Payloads/Serverbound/MlsCommitWelcomePayload.cs
@@ -1,10 +1,0 @@
-namespace DSharpPlus.Voice.Protocol.Dave.V1.Gateway.Payloads.Serverbound;
-
-/// <summary>
-/// Represents a payload to <see cref="VoiceGatewayOpcode.MlsCommitWelcome"/>.
-/// </summary>
-internal sealed record MlsCommitWelcomePayload : IMlsPayload
-{
-    /// <inheritdoc/>
-    public required byte[] MlsMessage { get; init; }
-}
